chore: private fns
